requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
django-cors-headers==4.3.1
Pillow==10.4.0
python-decouple==3.8
drf-spectacular==0.27.2
